Rename packet/header/* files for consistency.
