+++ src/packet/header/mod.rs
@@ -1,11 +1,11 @@
 mod header_parser;
 mod header_reader;
-mod packet;
-mod fragment;
-mod heart_beat;
+mod packet_header;
+mod fragment_header;
+mod heart_beat_header;
 
 pub use self::header_parser::HeaderParser;
 pub use self::header_reader::HeaderReader;
-pub use self::packet::PacketHeader;
-pub use self::fragment::FragmentHeader;
-pub use self::heart_beat::HeartBeatHeader;
+pub use self::packet_header::PacketHeader;
+pub use self::fragment_header::FragmentHeader;
+pub use self::heart_beat_header::HeartBeatHeader;
